added remanining tests;
int is failing because Unmarshal uses float64 for all number types

+++ internal/provider/flag_store.go
@@ -8,6 +8,6 @@ var IntFlagValues = make(map[string]IntFlagValue)
 func PopulateFlagValues() {
 	StringFlagValues["dataplane_generation"] = StringFlagValue{FlagKey: "dataplane_generation", FlagValue: "metal.v1"}
 	BoolFlagValues["grant_soil_access"] = BoolFlagValue{FlagKey: "grant_soil_access", FlagValue: false}
-	FloatFlagValues["special_ability_buff_perc"] = FloatFlagValue{FlagKey: "special_ability_buff_perc", FlagValue: 0.23456}
-	IntFlagValues["num_of_special_abilities"] = IntFlagValue{FlagKey: "num_of_special_abilities", FlagValue: 12}
+	FloatFlagValues["special_ability_buff_perc"] = FloatFlagValue{FlagKey: "special_ability_buff_perc", FlagValue: float64(0.23456)}
+	IntFlagValues["num_of_special_abilities"] = IntFlagValue{FlagKey: "num_of_special_abilities", FlagValue: int64(12)}
 }
